Change the sequence in which the camera is updated
The camera is now updated after the things "sprites" are drawn and
before the players "sprites" are drawn.

This solves a glitch were the things were drawn at an angle slightly not
facing the camera and another glitch were the camera was following the
player because it was late to its position.