Fix insane VRAM use by textures loaded multiple times
Blazing fast level loading, the game stops using hundreds of megs of RAM
on systems with low video memory. A bug caused some textures to be
loaded twice on some special occurrences. The old code was seeking
optimisation by placing every texture in a alphabetical order but it
didn't work as expected.