This fixes #19 (read the description)
Player sprites are stretched vertically, but keep the same ratio so that
their width is not constant like the ingame width of their hitbox.
Previously, the width of the sprite was calculated using the thing's
radius (radius*2), now it's calculated using a ratio between the
hitbox's height (to which the sprite is vertically stretched) and the
sprite's height.